apps/mofa-rss-newscaster/agents/rss-headline-scripter/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rss-headline-scripter"
version = "0.1.0"
description = "MoFA agent that converts RSS entries into headline scripts"
readme = "README.md"
license = {text = "MIT License"}
dependencies = [
    "numpy<2.0.0",
    "pyarrow>=5.0.0",
    "mofa-core",
    "dora-rs",
]

[project.scripts]
rss-headline-scripter = "rss_headline_scripter.main:main"

[tool.setuptools.packages.find]
include = ["rss_headline_scripter*"]
